Update config structure
