+++ .github/codeql/codeql-config.yml
@@ -1,0 +1,5 @@
+name: CodeQL Configuration
+
+paths-ignore:
+  - node_modules
+  - dist
