--- .github/linters/.shellcheckrc
@@ -1,1 +1,0 @@
-disable=SC2034,SC2181
