+++ .github/workflows/continuous-integration.yml
@@ -2,7 +2,7 @@ name: Continuous Integration
 
 on:
   push:
-    branches:
+    branches-ignore:
       - main
   pull_request:
     branches:
