+++ .github/workflows/super-linter.yml
@@ -2,7 +2,7 @@ name: Super Linter
 
 on:
   push:
-    branches:
+    branches-ignore:
       - main
   pull_request:
     branches:
