+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "parser",
   "description": "Convert issue form responses to JSON",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "author": "Nick Alteen <ncalteen@github.com>",
   "homepage": "https://github.com/issue-ops/parser#readme",
   "repository": {
